split Core Env file
