+++ compiler/Plato/Core/Data.hs
@@ -1,39 +1,10 @@
 module Plato.Core.Data where
-
-import Control.Monad.IO.Class
-import GHC.IORef
+ 
 
 import Plato.Common.Ident
 import Plato.Common.Pretty
-import Plato.Core.Scope
+import Plato.Core.Closure
 import Plato.Syntax.Core
-
-class Env e where
-        extendE :: MonadIO m => PrtInfo -> e -> m Index
-        getE :: MonadIO m => Index -> e -> m EnvEntry
-        setE :: MonadIO m => Index -> EnvEntry -> e -> m ()
-        prtE :: MonadIO m => Index -> e -> m PrtInfo
-
-set :: [a] -> Int -> a -> [a]
-set [] _ _ = error "list is empty"
-set (_ : as) 0 b = b : as
-set (a : as) i b = a : set as (i - 1) b
-
-instance Env (IORef EnvEntries) where
-        extendE fi ref = do
-                env <- liftIO $ readIORef ref
-                let i = length env
-                liftIO $ writeIORef ref (env ++ [(Index i, fi)])
-                return i
-        getE i ref = do
-                env <- liftIO $ readIORef ref
-                return $ fst $ env !! i
-        setE i v ref = do
-                env <- liftIO $ readIORef ref
-                liftIO $ writeIORef ref (set env i (v, snd (env !! i)))
-        prtE i ref = do
-                env <- liftIO $ readIORef ref
-                return $ snd $ env !! i
 
 newtype Boxed = Boxed (Clos Term) deriving (Eq, Show)
 
@@ -69,8 +40,6 @@ data EnvEntry
         = Index Index
         | Closure (Clos Term)
         deriving (Show)
-
-type EnvEntries = [(EnvEntry, PrtInfo)]
 
 data PrtInfo = PrtInfo
         { name :: Ident
