+++ compiler/Plato/Core/Pretty.hs
@@ -7,9 +7,10 @@ import Control.Monad.Reader.Class
 import Plato.Common.Ident
 import Plato.Common.Pretty
 import Plato.Common.Uniq
+import Plato.Core.Closure
 import Plato.Core.Data
+import Plato.Core.Env
 import Plato.Core.Normalise
-import Plato.Core.Scope
 import Plato.Syntax.Core
 
 class Print a where
